+++ R/int_genNegBinom.R
@@ -6,15 +6,22 @@
 # @param formula String that specifies the mean
 # @return A data.frame column with the updated simulated data
 
-.gennegbinom <- function(n, formula, dispersion, link="identity", dtSim) {
-
+.getNBmean <- function(dtSim, formula, link) {
+  
   if (link == "log") {
     logmean <- with(dtSim,eval(parse(text = as.character(formula))))
     mean <- exp(logmean)
   } else {
     mean <- with(dtSim,eval(parse(text = as.character(formula))))
   }
+  
+  return(mean)
+  
+}
 
+.gennegbinom <- function(n, formula, dispersion, link="identity", dtSim) {
+
+  mean <- .getNBmean(dtSim, formula, link)
   d <- as.numeric(as.character(dispersion))
 
   sp <- negbinomGetSizeProb(mean = mean, dispersion = d)