+++ NEWS.md
@@ -76,4 +76,4 @@
 
 # simstudy 0.1.11
 
-* Added negative binomial option to genCorGen, #addCorGen, genCorFlex, and addCorFlex#
+* Added negative binomial option to genCorGen, addCorGen, genCorFlex, and addCorFlex